# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkyReserve Flight Booking Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /passengers:
    get:
      operationId: listPassengers
      tags: [passengers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SavedPassenger" }
    post:
      operationId: createPassenger
      tags: [passengers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SavedPassengerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavedPassenger" }
  /passengers/{id}:
    patch:
      operationId: updatePassenger
      tags: [passengers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SavedPassengerInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavedPassenger" }
    delete:
      operationId: deletePassenger
      tags: [passengers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [paymentMethods]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PaymentMethod" }
    post:
      operationId: createPaymentMethod
      tags: [paymentMethods]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentMethodInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentMethod" }
  /payment-methods/{id}:
    delete:
      operationId: deletePaymentMethod
      tags: [paymentMethods]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /wishlist:
    get:
      operationId: listWishlist
      tags: [wishlist]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WishlistItem" }
    post:
      operationId: createWishlistItem
      tags: [wishlist]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WishlistInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WishlistItem" }
  /wishlist/{id}:
    delete:
      operationId: deleteWishlistItem
      tags: [wishlist]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Flight:
      type: object
      required: [id, airlineId, airlineName, departure, arrival, departureTime, arrivalTime, status, aircraft, totalSeats, availableSeats, basePrice, durationMinutes]
      properties:
        id: { type: integer }
        airlineId: { type: integer }
        airlineName: { type: string }
        flightNumber: { type: string }
        departure: { type: string }
        arrival: { type: string }
        departureTime: { type: string }
        arrivalTime: { type: string }
        status: { type: string }
        aircraft: { type: string }
        totalSeats: { type: integer }
        availableSeats: { type: integer }
        basePrice: { type: number }
        durationMinutes: { type: integer }
...
    SavedPassenger:
      type: object
      required: [id, customerId, fullName, email, phone]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        passportNumber: { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
    SavedPassengerInput:
      type: object
      required: [fullName, email, phone]
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        passportNumber: { type: string }
        dateOfBirth: { type: string }
    PaymentMethod:
      type: object
      required: [id, customerId, cardBrand, last4, expiryMonth, expiryYear, cardholderName]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        cardBrand: { type: string }
        last4: { type: string }
        expiryMonth: { type: integer }
        expiryYear: { type: integer }
        cardholderName: { type: string }
    PaymentMethodInput:
      type: object
      required: [cardBrand, cardNumber, expiryMonth, expiryYear, cardholderName]
      properties:
        cardBrand: { type: string, minLength: 1 }
        cardNumber: { type: string, minLength: 4 }
        expiryMonth: { type: integer }
        expiryYear: { type: integer }
        cardholderName: { type: string, minLength: 1 }
    WishlistItem:
      type: object
      required: [id, customerId, flight]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        flight: { $ref: "#/components/schemas/Flight" }
    WishlistInput:
      type: object
      required: [flightId]
      properties:
        flightId: { type: integer }
...