# lib/api-zod/src/generated/types/savedPassenger.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SkyReserve Flight Booking Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedPassenger {
  id: number;
  customerId: number;
  fullName: string;
  email: string;
  phone: string;
  /** @nullable */
  passportNumber?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
}